Finished API call class
Added a url.txt file which the SerpAPI.search(url); function writes the url parameter to. The function then runs the serpAPI.py file, which reads the url.txt file to make the api call, and saves the returned search results, to serpAPI.json, which is read, sorted, and organized by the getResults() function and then returned through the search function.